Performance improvements
  - Only handle events if they cause a change
 -  Cache the PixelMap in ColorWheel
